Fix some mail actions e.g. "view as source" or "Html" no longer working

+++ api/src/Header/UserAgent.php
@@ -32,11 +32,11 @@ class UserAgent
 	/**
 	 * Version of user-agent as specified by browser
 	 *
-	 * @return string
+	 * @return integer
 	 */
 	public static function version()
 	{
-		return self::$ua_version;
+		return (int)self::$ua_version;
 	}
 
 	/**
